feat: include file from path
previously, when @Fn::IncludeFile is embedded in a file that is itself included by a @Fn::IncludeFile at a higher level, and the arguments of the nested @Fn::IncludeFile depend on an attribute that is not yet available, the nested @Fn::IncludeFile forgets the path of the file it came from. To account for that, the nested @Fn::IncludeFile will be rewritten as @Fn::IncludeFileFromPath, and the current path will be prepended to the arguments so that pathing is preserved.